implementate modifiche generazione studeid con data

+++ CoreTestWebApp/CoreTestWebApp/Repositories/IStudentIdGenerator.cs
@@ -4,6 +4,6 @@ namespace CoreTestWebApp.Repositories
 {
     public interface IStudentIdGenerator
     {
-        int GenerateIdForStudent(Student student);  
+        long GenerateIdForStudent(Student student);     
     }
 }
